Fix bug of failing documentation builds when not tags are present in Git repository
In the sphinx config file one can specify, what tags and branches to build the documentation of. When the selection is empty (e.g. not tags in the repository) the build process throws an exception. This is now fixed and the prcess either exits gracefully or buids the development version of documentation, depending on the used command line flags